% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainwritR-package.R
\docType{package}
\name{brainwritR-package}
\alias{brainwritR}
\alias{brainwritR-package}
\title{brainwritR: resilient classroom brainwriting}
\description{
A single-process, SQLite-backed German-language Shiny application for
facilitated topic rotation. See \code{\link[=run_app]{run_app()}} and the classroom guide.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CTTIR/brainwritR}
  \item \url{https://cttir.github.io/brainwritR/}
  \item Report bugs at \url{https://github.com/CTTIR/brainwritR/issues}
}

}
\author{
\strong{Maintainer}: R. Heller \email{raban.heller@uni-ulm.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item R. Heller \email{raban.heller@uni-ulm.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID}) [copyright holder]
}

}
